feat: worked on comments by Martjin
