+++ packages/vendure-plugin-payment-extensions/eslint.config.js
@@ -1,0 +1,13 @@
+module.exports = (async () => {
+  const { default: parentConfig } = await import('../../eslint-base.config.js');
+  return [
+    ...parentConfig,
+    {
+      languageOptions: {
+        parserOptions: {
+          project: './tsconfig.json',
+        },
+      },
+    },
+  ];
+})();
