+++ packages/vendure-plugin-payment-extensions/package.json
@@ -22,6 +22,6 @@
     "start": "ts-node test/dev-server.ts",
     "build": "rimraf dist && tsc && copyfiles -u 1 'src/ui/**/*' dist/",
     "test": "vitest run",
-    "lint": "echo 'No linting configured'"
+    "lint": "eslint ."
   }
 }
