+++ packages/vendure-plugin-payment-extensions/src/index.ts
@@ -1,3 +1,3 @@
 export * from './payment-extensions-plugin';
-export * from './settle-without-payment-checker';
+export * from './is-customer-In-group-payment-checker';
 export * from './settle-without-payment-handler';
