+++ packages/vendure-plugin-payment-extensions/src/settle-without-payment-handler.ts
@@ -4,22 +4,22 @@ import { LanguageCode, Logger, PaymentMethodHandler } from '@vendure/core';
  * Marks an order as placed, without any payment needed. This is used for purchase orders, where customers need to be in a specific group.
  */
 export const settleWithoutPaymentHandler = new PaymentMethodHandler({
-  code: 'settle-without-payment-handler',
+  code: 'settle-order-without-payment',
   description: [
     {
       languageCode: LanguageCode.en,
-      value: 'Settle Without Payment Handler',
+      value: 'Settle order without payment',
     },
   ],
   args: {},
-  createPayment: (ctx, order, amount, args, metadata) => {
+  createPayment: (_, order, amount, __, metadata) => {
     Logger.info(
-      `'${order.code}' settled without an upfront pay for customer '${order.customer?.emailAddress}'`
+      `'${order.code}' settled without payment for customer '${order.customer?.emailAddress}'`
     );
     return {
       amount,
       state: 'Settled',
-      metadata: { note: 'Settled without payment' },
+      metadata: { ...metadata, note: 'Settled without payment' },
     };
   },
   settlePayment: () => ({
